Recieve the boot command so that we can reschedule alarms. Centralize the scheduling alarm code
